rename: 체크스타일 해결
